Return every value in its column's declared type, and type bitwise results as ACE does
A CASE, IIF, COALESCE, GREATEST or LEAST column and a set operation's
column now declare one type and convert each value to it, so the type a
reader reports is the type it returns.

- CASE, IIF, COALESCE, GREATEST and LEAST widen numbers on one ladder:
  the wider whole number, a Single with a Byte or an Integer, a Double
  for a Single with anything wider or with a Double, a Decimal with a
  whole number, and a Double for Currency with a Large Number. IIF no
  longer declares nothing when its branches differ.
- UNION, INTERSECT and EXCEPT type each column from both queries, as ACE
  does: the same numeric ladder with a Boolean as the Integer -1 or 0; a
  bare NULL takes the other query's type; a GUID or binary value with
  anything else makes a binary column of each value's bytes; any other
  mix is text. Values are converted before rows are compared.
- BNOT of a Boolean or an Integer, and BAND, BOR and BXOR of two of them,
  give an Integer; anything else gives a Long, or an Int64.

+++ test/LibRed.Engine.Tests/ProgramFlowFunctionTests.cs
@@ -47,9 +47,9 @@ public class ProgramFlowFunctionTests(ProgramFlowFunctionTests.Database database
         Assert.Equal(expected, Scalar($"IIF({condition}, 'T', 'F')"));
 
     [Theory]
-    [InlineData("IIF(TRUE, 1, 1/0)", 1)]
-    [InlineData("IIF(FALSE, 1/0, 2)", 2)]
-    public void Iif_evaluates_only_the_branch_it_takes(string expression, int expected) =>
+    [InlineData("IIF(TRUE, 1, 1/0)", 1.0)]
+    [InlineData("IIF(FALSE, 1/0, 2)", 2.0)]
+    public void Iif_evaluates_only_the_branch_it_takes(string expression, double expected) =>
         Assert.Equal(expected, Scalar(expression));
 
     [Theory]
